backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-core
langchain-openai
pydantic
python-multipart

# Auth + DB
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]

pandas
openpyxl
xlrd

PyPDF2
python-docx

Pillow
pytesseract

requests

